+++ flake.nix
@@ -3,23 +3,26 @@
 
   inputs = {
     nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";
+
     home-manager = {
       url = "github:nix-community/home-manager";
-      inputs.nixpkgs.follows = "nixpkgs";
+    #   inputs.nixpkgs.follows = "nixpkgs";
     };
+
     noctalia = {
       url = "github:noctalia-dev/noctalia-shell";
-      inputs.nixpkgs.follows = "nixpkgs";
+      #inputs.nixpkgs.follows = "nixpkgs";
     };
+
     niri-flake = {
       url = "github:sodiboo/niri-flake";
-      inputs.nixpkgs.follows = "nixpkgs";
     };
-    # Add NixVim input
+
     nixvim = {
       url = "github:nix-community/nixvim";
-      inputs.nixpkgs.follows = "nixpkgs";
+      # removed follows to fix version mismatch
     };
+
   };
 
   nixConfig = {
@@ -48,7 +51,6 @@
           ./home/home.nix
           noctalia.homeModules.default
           niri-flake.homeModules.niri
-          # Add NixVim module for home-manager
           nixvim.homeModules.nixvim
         ];
       };