+++ system/modules/appimage.nix
@@ -1,0 +1,6 @@
+{ config, pkgs, ... }: {
+  programs.appimage = {
+    enable = true;
+    binfmt = true;
+  };
+}